rename mine model and typo
