+++ api/src/models/bcmi/index.js
@@ -1,1 +1,1 @@
-exports.mineBCMI= require('./mine-bcmi');
+exports.mineBCMI= require('./mine');
